Create HTMl file with intial boilerplate
'index.html' will contain the HTML code which the website will be built
with. A boilerplate has been implemented, this includes the HTML version
being defined as the latest version, HTML5, the language defined as
English. The character encoding method for the document has been
specified as UTF-8 and 'Odin Recipes' will appear as both a title in the
browser tab and the main heading of the webpage.